+++ app/serializers/guild_base_serializer.rb
@@ -1,0 +1,3 @@
+class GuildBaseSerializer < ActiveModel::Serializer
+  attributes :id, :full_name
+end