+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Modify
 - requests to DB to find available events
+- rebuild event calendar for faster filtering
 
 ## [0.8.5] - 2019-11-21
 ### Added